Rework test
